refactor: 💡 rename module components
