+++ v2/src/modules/tokens/components/TokenDetailsHolder/TableTokenHolderRow.vue
