+++ v2/src/modules/tokens/components/TokenDetailsTransfer/TableTokenTransferRow.vue
