+++ v2/src/modules/tokens/components/TokenMarketInfo/TableTokenMarketInfoRow.vue
