+++ v2/src/modules/txs/components/TableTxsRow.vue
